requirements.txt
deap==1.4.2
editdistance==0.8.1
ga4qc==0.3.0
joblib==1.4.2
llvmlite==0.44.0
numba==0.61.0
numpy==2.1.0
pynndescent==0.5.13
quasim==1.0.0
scikit-learn==1.6.1
scipy==1.15.1
threadpoolctl==3.6.0
tqdm==4.67.1
umap-learn==0.5.7
